A LOT of things. Read Below
Separated the css of the individual items to a single document for
cleanliness. Changed the Subject of the email that will be sent if the
item is reserved to the name of the item. Erased some documents that
aren’t used. Tried to eliminate unnecessary comments. Fixed some links
that didn’t work due to renamed documents or private characters like
“backlash”.